feat(cli): update get object ref params
